# api/openapi.yaml
openapi: 3.0.3
info:
  title: Orbit
  description: Orbit API documentation.
  version: 1.0.0
servers:
  - url: http://localhost:9132
    description: Local Development Server
  - url: https://Orbit.Akarta.tech
    description: Production Server
paths:
...
  /api/auth/signup:
    post:
      summary: Sign up user
      description: User can signup here either with a seller, buyer, or admin profile.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - firstName
                - lastName
                - emailId
                - password
                - age
                - role
              properties:
                firstName:
                  type: string
                  example: "Aayush"
                lastName:
                  type: string
                  example: "Verma"
                emailId:
                  type: string
                  example: "aayush@email.com"
                password:
                  type: string
                  example: "SecuredPassword@123"
                age:
                  type: integer
                  example: 25
                role:
                  type: string
                  enum: [seller, buyer, admin]
                  example: "seller"
      responses:
        "201":
          description: Register Successfully (Returns user object and sets HttpOnly cookie)
          headers:
            Set-Cookie:
              description: Contains the authentication token
              schema:
                type: string
                example: "token=eyJhbGciOiJIUz...; Path=/; Max-Age=7200; HttpOnly"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Register Successfully"
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "60d5ec49f83b2c3a54cd5291"
                      firstName:
                        type: string
                        example: "Aayush"
                      lastName:
                        type: string
                        example: "Verma"
                      emailId:
                        type: string
                        example: "aayush@email.com"
                      age:
                        type: integer
                        example: 25
                      role:
                        type: string
                        example: "seller"
        "400":
          description: Bad Request (Validation failed or user already exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Key: 'RegisterRequest.EmailId' Error:Field validation for 'EmailId' failed on the 'email' tag"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal Server Error"
...